Feat/prep staging migration (#114)
* remove old migration

* add deprecated exception

* add proxy class builder to source model

* use source proxy builder

* make land_id not null

* lint import

* add multy depts options

* rename departement options

* add departements filtering

* update mep with all filtering options

* use base_qs to limit epci links creation

* move load class into models

* fixes

* rename file

* move factory to factories file and slice specific rules into command

* fix self referency
