+++ project/management/commands/build_project_ocsge_status.py
@@ -11,11 +11,26 @@ logger = logging.getLogger("management.commands")
 class Command(BaseCommand):
     help = "Build all ocsge status of projects"
 
+    def add_arguments(self, parser):
+        parser.add_argument(
+            "--departements",
+            nargs="+",
+            type=int,
+            help="Select departements to build",
+        )
+
     def handle(self, *args, **options):
+        logger.info("Start building ocsge status")
+
         projects = Project.objects.all()
+
+        if options.get("departements"):
+            logger.info("Filtering on departements: %s", options["departements"])
+            projects = projects.filter(cities__departement__source_id__in=options["departements"])
+
         count = projects.count()
 
-        logger.info(f"Start building ocsge status for {count} projects")
+        logger.info(f"{count} projects")
 
         for i, project in enumerate(projects):
             logger.info(f"{i + 1}/{count} - Process project {project.id}")
