+++ public_data/management/commands/build_administrative_layers.py
@@ -3,6 +3,7 @@ import logging
 from django.contrib.gis.db.models import Union
 from django.core.management.base import BaseCommand
 from django.core.paginator import Paginator
+from django.db.models import QuerySet
 
 from public_data.models import Cerema, Commune, Departement, Epci, Region
 from public_data.models.administration import Scot
@@ -20,8 +21,16 @@ class Command(BaseCommand):
             action="store_true",
             help="Clean all data before loading",
         )
+        parser.add_argument(
+            "--departements",
+            nargs="+",
+            type=int,
+            help="Select departements to build",
+        )
 
     def handle(self, *args, **options):
+        """This command is keeped for documentation prupose, do not use it unless to be sure of review everything."""
+
         clean = options.get("clean", False)
 
         logger.info("Recreate region, departement, EPCI and communes referentials")
@@ -34,16 +43,22 @@ class Command(BaseCommand):
             Commune.objects.all().delete()
             Scot.objects.all().delete()
 
-        self.load_region()
-        self.load_departement()
-        self.load_epci()
-        self.link_epci()
-        self.load_communes(table_was_cleaned=clean)
+        base_qs = Cerema.objects.all()
 
-    def load_region(self):
+        if options.get("departements"):
+            base_qs = base_qs.filter(dept_id__in=options["departements"])
+
+        self.load_region(base_qs)
+        self.load_departement(base_qs)
+        self.load_epci(base_qs)
+        self.load_scot(base_qs)
+        self.link_epci(base_qs)
+        self.load_communes(base_qs, table_was_cleaned=clean)
+
+    def load_region(self, base_qs: QuerySet):
         logger.info("Loading regions")
 
-        qs = Cerema.objects.values("region_id", "region_name", "srid_source")
+        qs = base_qs.values("region_id", "region_name", "srid_source")
         qs = qs.annotate(mpoly=Union("mpoly")).order_by("region_name")
 
         logger.info("%d found regions", len(qs))
@@ -66,12 +81,12 @@ class Command(BaseCommand):
         logger.info("%d regions created", total_created)
         logger.info("Done loading regions")
 
-    def load_departement(self):
+    def load_departement(self, base_qs: QuerySet):
         logger.info("Loading departements")
 
         regions = {r.source_id: r for r in Region.objects.all()}
 
-        qs = Cerema.objects.values("region_id", "dept_id", "dept_name", "srid_source")
+        qs = base_qs.values("region_id", "dept_id", "dept_name", "srid_source")
         qs = qs.annotate(mpoly=Union("mpoly")).order_by("dept_id")
 
         logger.info("%d departements found", len(qs))
@@ -95,10 +110,10 @@ class Command(BaseCommand):
         logger.info("%d departements created", total_created)
         logger.info("Done loading departements")
 
-    def load_epci(self):
+    def load_epci(self, base_qs: QuerySet):
         logger.info("Loading EPCI")
 
-        qs = Cerema.objects.values("epci_id", "epci_name", "srid_source")
+        qs = base_qs.values("epci_id", "epci_name", "srid_source")
         qs = qs.annotate(mpoly=Union("mpoly")).order_by("epci_id")
 
         logger.info("%d EPCI found", len(qs))
@@ -121,10 +136,10 @@ class Command(BaseCommand):
         logger.info("%d EPCI created", total_created)
         logger.info("Done loading EPCI")
 
-    def load_scot(self):
+    def load_scot(self, base_qs: QuerySet):
         logger.info("Loading SCOT")
 
-        qs = Cerema.objects.values("scot", "srid_source")
+        qs = base_qs.values("scot", "srid_source")
         qs = qs.annotate(mpoly=Union("mpoly")).order_by("scot")
 
         logger.info("%d SCoTs found", len(qs))
@@ -168,23 +183,23 @@ class Command(BaseCommand):
             scot.departements.add(*[depts[d] for d in data["departement_ids"]])
             scot.regions.add(*[regions[r] for r in data["region_ids"]])
 
-    def link_epci(self):
+    def link_epci(self, base_qs: QuerySet):
         logger.info("Link EPCI <-> département")
         depts = {d.source_id: d for d in Departement.objects.all()}
         epcis = {e.source_id: e for e in Epci.objects.all()}
         for epci in epcis.values():
             epci.departements.remove()
-        links = Cerema.objects.values_list("epci_id", "dept_id").distinct()
+        links = base_qs.values_list("epci_id", "dept_id").distinct()
         logger.info("%d links found", links.count())
         for epci_id, dept_id in links:
             epcis[epci_id].departements.add(depts[dept_id])
         logger.info("Done linking")
 
-    def load_communes(self, table_was_cleaned: bool):
+    def load_communes(self, base_qs: QuerySet, table_was_cleaned: bool):
         logger.info("Loading Communes")
         depts = {d.source_id: d for d in Departement.objects.all()}
         epcis = {e.source_id: e for e in Epci.objects.all()}
-        qs = Cerema.objects.all().order_by("city_insee")
+        qs = base_qs.order_by("city_insee")
         paginator = Paginator(qs, 1000)
         logger.info("%d Communes found ", qs.count())
