+++ public_data/management/commands/mep_45.py
@@ -16,7 +16,6 @@ class Command(BaseCommand):
 
     def load_departement(self, departement: Departement):
         call_command("load_ocsge", departement=departement.name)
-        call_command("setup_departements")
         call_command("build_commune_data", departement=departement.name, verbose=True)
         call_command("build_artificial_area", departement=departement.name, verbose=True)
         call_command("import_gpu", departement=departement.name)
@@ -26,22 +25,23 @@ class Command(BaseCommand):
 
         call_command("maintenance", on=True)
 
+        logger.info("Initialize data sources")
         call_command("loaddata", "public_data/models/data_source_fixture.json")
 
-        call_command("load_cerema", departement="971")  # Guadeloupe
-        call_command("load_cerema", departement="972")  # Martinique
-        call_command("load_cerema", departement="973")  # Guyane française
-        call_command("load_cerema", departement="974")  # La Réunion
-
-        call_command("build_administrative_layers")
-        call_command("load_scot")
+        logger.info("Load DROM-COM")
+        # Guadeloupe, Martinique, Guyane française, La Réunion
+        dept_codes = ["971", "972", "973", "974"]
+        call_command("load_cerema", official_land_ids=dept_codes)
+        call_command("build_administrative_layers", departements=dept_codes)
         call_command("load_insee")
 
+        logger.info("Load new OCS GE")
+        call_command("setup_departements")
         self.load_departement(departement=Departement.objects.get(name="Hauts-de-Seine"))
         self.load_departement(departement=Departement.objects.get(name="Landes"))
 
-        call_command("reset_first_last")
-        call_command("build_project_ocsge_status")
+        call_command("reset_first_last", departements=dept_codes)
+        call_command("build_project_ocsge_status", departements=dept_codes)
 
         call_command("maintenance", off=True)
         logger.info("End mep_45")
