+++ public_data/models/__init__.py
@@ -1,7 +1,7 @@
 from .administration import *  # noqa: F401, F403
 from .couverture_usage import *  # noqa: F401, F403
+from .data_source import *  # noqa: F401, F403
 from .deprecated import *  # noqa: F401, F403
 from .gpu import *  # noqa: F401, F403
 from .mixins import *  # noqa: F401, F403
 from .ocsge import *  # noqa: F401, F403
-from .source import *  # noqa: F401, F403
