+++ public_data/models/couverture_usage.py
@@ -2,6 +2,8 @@
 Ce fichier contient les référentiels CouvertureSol et UsageSol qui sont les deux
 types d'analyse fournies par l'OCSGE.
 """
+from functools import cache
+
 from django.db import models
 
 
@@ -210,6 +212,7 @@ class CouvertureUsageMatrix(models.Model):
         return f"{cs}-{us}:{a}{c}{n}"
 
     @classmethod
+    @cache
     def matrix_dict(cls):
         _matrix_dict = dict()
